+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Move slides property to getter on `ProductImages` component.
+
+### Fixed
+- Fix `forceUpdate` being called on instantiation of the `ProductImage` component.
 
 ## [3.12.0] - 2019-01-30
 ### Changed